+++ src/Provider/GoogleMapsPlaces/Tests/IntegrationTest.php
@@ -1,0 +1,46 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Geocoder package.
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * @license    MIT License
+ */
+
+namespace Geocoder\Provider\GoogleMapsPlaces\Tests;
+
+use Geocoder\IntegrationTest\ProviderIntegrationTest;
+use Geocoder\Provider\GoogleMapsPlaces\GoogleMapsPlaces;
+use Http\Client\HttpClient;
+
+/**
+ * @author Tobias Nyholm <tobias.nyholm@gmail.com>
+ */
+class IntegrationTest extends ProviderIntegrationTest
+{
+    protected $testAddress = false;
+
+    protected $testReverse = false;
+
+    protected $testIpv4 = false;
+
+    protected $testIpv6 = false;
+
+    protected function createProvider(HttpClient $httpClient)
+    {
+        return new GoogleMapsPlaces($httpClient, $_SERVER['GOOGLE_GEOCODING_KEY']);
+    }
+
+    protected function getCacheDir()
+    {
+        return __DIR__.'/.cached_responses';
+    }
+
+    protected function getApiKey()
+    {
+        return $_SERVER['GOOGLE_GEOCODING_KEY'];
+    }
+}