+++ src/Provider/Nominatim/Nominatim.php
@@ -230,9 +230,14 @@ final class Nominatim extends AbstractHttpProvider implements Provider
 
         $location = $builder->build(NominatimAddress::class);
         $location = $location->withAttribution($place->licence);
-        $location = $location->withOSMId(intval($place->osm_id));
-        $location = $location->withOSMType($place->osm_type);
         $location = $location->withDisplayName($place->display_name);
+
+        if (isset($place->osm_id)) {
+            $location = $location->withOSMId(intval($place->osm_id));
+        }
+        if (isset($place->osm_type)) {
+            $location = $location->withOSMType($place->osm_type);
+        }
 
         if (false === $reverse) {
             $location = $location->withCategory($place->category);