+++ src/Provider/Chain/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 The change log describes what is "Added", "Removed", "Changed" or "Fixed" between each release.
 
+## 4.1.1
+
+### Fixed
+
+- Fix (and improve) error message when provider can not **reverse** geocode
+
+## 4.1.0
+
+### Added
+
+- Add configurable log levels
+
 ## 4.0.1
 
 ### Fixed